itsmeSimple using inheritance
Inheritance applied in itsmeSimple.py
